requirements.txt
anyio==3.7.1
fastapi[all]>=0.115.3
fastapi-health
python-dotenv
sqlmodel
toolz
pysaml2
pytest
starlette-context
casbin
pytest
canvasapi
httpx
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-text-splitters
langchain-openai
langsmith
loguru
openai
pandas
psycopg2
pydantic
PyJWT
PyMuPDF
PyMuPDFb
requests
rsa
SQLAlchemy
uvicorn
tiktoken
pypdf
python-magic
langchain-huggingface

